[core] -Fix: hid_attrib: copying HATT_INTEGER is really copying long *
git-svn-id: svn://repo.hu/librnd/trunk@32561 1d61de99-560f-4dba-8957-1e505d42358b

+++ src/librnd/core/hid_attrib.c
@@ -104,7 +104,7 @@ int rnd_hid_parse_command_line(int *argc, char ***argv)
 				if ((backup != NULL) && (a->value == NULL))
 					a->value = &backup[i].lng;
 				if (a->value)
-					*(int *) a->value = a->default_val.lng;
+					*(long *) a->value = a->default_val.lng;
 				break;
 			case RND_HATT_COORD:
 				if ((backup != NULL) && (a->value == NULL))
@@ -171,7 +171,7 @@ int rnd_hid_parse_command_line(int *argc, char ***argv)
 						break;
 					case RND_HATT_INTEGER:
 						if (a->value)
-							*(int *) a->value = strtol((*argv)[1], 0, 0);
+							*(long *) a->value = strtol((*argv)[1], 0, 0);
 						else
 							a->default_val.lng = strtol((*argv)[1], 0, 0);
 						(*argc)--;
